Materialに任意の定数バッファーを渡せるようになったよ MaterialIDをu64からusizeに変更。この変更はあまり意味が無かったかもしれない。 shapell_shader_PS.rsを追加。シェーダーをオブジェクトに合わせて切り替えよう。 antenna_chain.rs、controller.rsのテストケースを修正。tsumugiSimpleHashMapを作成。HashMapの高速探査とVecのメモリ連続性を併せ持ちたい。これをマテリアル保管場所に指定。定数バッファーを作成するにあたってCpID3D12Resourceにリソースの実態や渡したいルートパラメータインデックスを渡せるようになった。これに伴い、構造体や関数を全面的に修正。これによりResourceを扱いやすくなったかも。fetch_materialdataにマテリアルの更新が行えるように機能追加。同じマテリアルを認識できないことによる、ディスクリプタヒープを使い切ってしまう問題を修正。定数バッファーをコマンドラインに流すように修正。D3D12_INPUT_ELEMENT_DESCに渡すSemanticNameについて、終端文字をrustが付けてくれないことを失念していたので、終端文字を追加。

+++ tsumugiShaderStock/src/lib.rs
@@ -21,7 +21,7 @@ pub struct TsumugiMaterial {
     ///マテリアルの名前が重複していた場合は古いマテリアルが消される。
     pub material:Material,
     ///マテリアルの固有の番号。これが違うと異なるマテリアルと認識される
-    pub material_element_id:u64,
+    pub material_element_id:usize,
     ///マテリアルの名前。
     pub material_name: &'static str,
 }
